backend/task1/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg==3.1.18
python-dotenv==1.0.0
dj-database-url==2.1.0
